+++ higan/component/processor/z80/z80.hpp
@@ -51,8 +51,8 @@ struct Z80 {
   auto displace(uint16&) -> uint16;
   auto read(uint16 addr) -> uint8;
   auto write(uint16 addr, uint8 data) -> void;
-  auto in(uint8 addr) -> uint8;
-  auto out(uint8 addr, uint8 data) -> void;
+  auto in(uint16 addr) -> uint8;
+  auto out(uint16 addr, uint8 data) -> void;
 
   //instruction.cpp
   auto instruction() -> void;
